+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Updated alertmanager dashboard to show related logs
+- Add loki mixins dashboards.
 
 ## [3.23.0] - 2024-08-22